[STEP17] added login helper and sessions spec
