+++ myapp/spec/rails_helper.rb
@@ -22,7 +22,7 @@ require 'rspec/rails'
 # directory. Alternatively, in the individual `*_spec.rb` files, manually
 # require only the support files necessary.
 #
-# Dir[Rails.root.join('spec', 'support', '**', '*.rb')].sort.each { |f| require f }
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].sort.each { |f| require f }
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.
@@ -66,6 +66,9 @@ RSpec.configure do |config|
 
   # adding FactoryBot
   config.include FactoryBot::Syntax::Methods
+
+  # LoginHelper
+  config.include LoginHelper, type: :system
 
   # Capybara config
