+++ myapp/spec/system/tasks_spec.rb
@@ -4,12 +4,14 @@ require 'rails_helper'
 
 RSpec.describe 'Tasks (System)', type: :system do
   let!(:user1) { create(:user, email: 'admin@tm.com', role: 1, password_digest: '$2a$12$NPcgfEO8vN91/zbwM5KwP.9NsKMVtEorU/Lk9tOw4SBxfMQ2tsjOO', name: 'admin') }
+  before do
+    login(user1)
+  end
   let!(:task1) { create(:task, name: 'task1', status: 1, priority: 0, label: 'aaa', due_date: Faker::Time.forward(days: 1, period: :evening), user_id: user1.id) }
   let!(:task2) { create(:task, name: 'task2', status: 1, priority: 0, label: 'aaa', due_date: Faker::Time.forward(days: 2, period: :evening), user_id: user1.id) }
   let!(:task3) { create(:task, name: 'task3', status: 2, priority: 0, label: 'aaa', due_date: Faker::Time.forward(days: 3, period: :evening), user_id: user1.id) }
   context 'CRUDing task' do
     it 'Add new task' do
-      login(user1)
       visit tasks_path
       click_on(class: 'newtask')
       fill_in 'task[name]', with: 'task4'
@@ -25,14 +27,12 @@ RSpec.describe 'Tasks (System)', type: :system do
 
   context 'Searching with status' do
     it 'Record found' do
-      login(user1)
       tasks = Task.ransack({ name_cont: nil, status_eq: 2 }).result
       expect(tasks.count).to eq 1
       expect(tasks).to include(task3)
     end
 
     it 'Record not found' do
-      login(user1)
       expect(Task.ransack({ name_cont: nil, status_eq: 0 }).result.count).to eq 0
     end
   end
@@ -59,16 +59,5 @@ RSpec.describe 'Tasks (System)', type: :system do
     it 'is record not found' do
       expect(Task.ransack({ name_cont: 'task4', status_eq: 2 }).result.count).to eq 0
     end
-  end
-
-  private
-
-  #TODO: fix this
-  def login(user)
-    visit login_path
-    fill_in 'session[email]', with: user.email
-    fill_in 'session[password]', with: 'abc1234'
-    click_button('Login')
-    expect(page).to have_text user.name
   end
 end
